Cardano.Node.CLI library
A module which contains common CLI utils:
* 'Protocol' definition
* and various optparse-applicative parsers
