+++ app/CLI.hs
@@ -19,57 +19,16 @@ module CLI (
   , progDesc
   ) where
 
-import           Data.Foldable (asum)
 import           Data.Semigroup ((<>))
 import           Options.Applicative
 
 import           Ouroboros.Consensus.BlockchainTime
-import           Ouroboros.Consensus.Demo
-import           Ouroboros.Consensus.Demo.Run
 import qualified Ouroboros.Consensus.Ledger.Mock as Mock
-import           Ouroboros.Consensus.NodeId (NodeId (..))
-import           Ouroboros.Consensus.Util
 
 import           TxSubmission (command', parseMockTx)
 import           Topology (TopologyInfo (..), NodeAddress (..))
 
-import qualified Test.Cardano.Chain.Genesis.Dummy as Dummy
-
-{-------------------------------------------------------------------------------
-  Untyped/typed protocol boundary
--------------------------------------------------------------------------------}
-
-data Protocol =
-    BFT
-  | Praos
-  | MockPBFT
-  | RealPBFT
-
-data SomeProtocol where
-  SomeProtocol :: RunDemo blk => DemoProtocol blk -> SomeProtocol
-
-fromProtocol :: Protocol -> IO SomeProtocol
-fromProtocol BFT =
-    case runDemo p of
-      Dict -> return $ SomeProtocol p
-  where
-    p = DemoBFT defaultSecurityParam
-fromProtocol Praos =
-    case runDemo p of
-      Dict -> return $ SomeProtocol p
-  where
-    p = DemoPraos defaultDemoPraosParams
-fromProtocol MockPBFT =
-    case runDemo p of
-      Dict -> return $ SomeProtocol p
-  where
-    p = DemoMockPBFT defaultDemoPBftParams
-fromProtocol RealPBFT =
-    case runDemo p of
-      Dict -> return $ SomeProtocol p
-  where
-    p = DemoRealPBFT defaultDemoPBftParams genesisConfig
-    genesisConfig = Dummy.dummyConfig
+import           Cardano.Node.CLI
 
 {-------------------------------------------------------------------------------
   Command line arguments
@@ -91,42 +50,6 @@ nodeParser = NodeCLIArguments
     <*> parseSlotDuration
     <*> parseCommand
 
-parseSystemStart :: Parser SystemStart
-parseSystemStart = option (SystemStart <$> auto) $ mconcat [
-      long "system-start"
-    , help "The start time of the system (e.g. \"2018-12-10 15:58:06\""
-    ]
-
-parseSlotDuration :: Parser SlotLength
-parseSlotDuration = option (mkSlotLength <$> auto) $ mconcat [
-      long "slot-duration"
-    , value (mkSlotLength 5)
-    , help "The slot duration (seconds)"
-    ]
-  where
-    mkSlotLength :: Integer -> SlotLength
-    mkSlotLength = slotLengthFromMillisec . (* 1000)
-
-parseProtocol :: Parser Protocol
-parseProtocol = asum [
-      flag' BFT $ mconcat [
-          long "bft"
-        , help "Use the BFT consensus algorithm"
-        ]
-    , flag' Praos $ mconcat [
-          long "praos"
-        , help "Use the Praos consensus algorithm"
-        ]
-    , flag' MockPBFT $ mconcat [
-          long "mock-pbft"
-        , help "Use the Permissive BFT consensus algorithm using a mock ledger"
-        ]
-    , flag' RealPBFT $ mconcat [
-          long "real-pbft"
-        , help "Use the Permissive BFT consensus algorithm using the real ledger"
-        ]
-    ]
-
 parseCommand :: Parser Command
 parseCommand = subparser $ mconcat [
     command' "node" "Run a node." $
@@ -134,15 +57,6 @@ parseCommand = subparser $ mconcat [
   , command' "submit" "Submit a transaction." $
       TxSubmitter <$> parseTopologyInfo <*> parseMockTx <*> parseProtocol
   ]
-
-parseNodeId :: Parser NodeId
-parseNodeId =
-    option (fmap CoreId auto) (
-            long "node-id"
-         <> short 'n'
-         <> metavar "NODE-ID"
-         <> help "The ID for this node"
-    )
 
 parseHostName :: Parser String
 parseHostName =
