+++ cardano-node/Cardano/Node/CLI.hs
@@ -1,0 +1,133 @@
+{-# LANGUAGE GADTs #-}
+
+module Cardano.Node.CLI (
+  -- * Untyped/typed protocol boundary
+    Protocol(..)
+  , SomeProtocol(..)
+  , fromProtocol
+  -- * Parsers
+  , parseSystemStart
+  , parseSlotDuration
+  , parseProtocol
+  , parseNodeId
+  , parseCoreNodeId
+  , parseNumCoreNodes
+  ) where
+
+import           Prelude
+
+import           Data.Foldable (asum)
+import           Data.Semigroup ((<>))
+import           Options.Applicative
+
+import           Ouroboros.Consensus.BlockchainTime
+import           Ouroboros.Consensus.Demo
+import           Ouroboros.Consensus.Demo.Run
+import           Ouroboros.Consensus.NodeId (NodeId (..), CoreNodeId (..))
+import           Ouroboros.Consensus.Util
+
+import qualified Test.Cardano.Chain.Genesis.Dummy as Dummy
+
+
+{-------------------------------------------------------------------------------
+  Untyped/typed protocol boundary
+-------------------------------------------------------------------------------}
+
+data Protocol =
+    BFT
+  | Praos
+  | MockPBFT
+  | RealPBFT
+
+data SomeProtocol where
+  SomeProtocol :: RunDemo blk => DemoProtocol blk -> SomeProtocol
+
+fromProtocol :: Protocol -> IO SomeProtocol
+fromProtocol BFT =
+    case runDemo p of
+      Dict -> return $ SomeProtocol p
+  where
+    p = DemoBFT defaultSecurityParam
+fromProtocol Praos =
+    case runDemo p of
+      Dict -> return $ SomeProtocol p
+  where
+    p = DemoPraos defaultDemoPraosParams
+fromProtocol MockPBFT =
+    case runDemo p of
+      Dict -> return $ SomeProtocol p
+  where
+    p = DemoMockPBFT defaultDemoPBftParams
+fromProtocol RealPBFT =
+    case runDemo p of
+      Dict -> return $ SomeProtocol p
+  where
+    p = DemoRealPBFT defaultDemoPBftParams genesisConfig
+    genesisConfig = Dummy.dummyConfig
+
+{-------------------------------------------------------------------------------
+  Command parsers
+-------------------------------------------------------------------------------}
+
+parseSystemStart :: Parser SystemStart
+parseSystemStart = option (SystemStart <$> auto) $ mconcat [
+      long "system-start"
+    , help "The start time of the system (e.g. \"2018-12-10 15:58:06\""
+    ]
+
+parseSlotDuration :: Parser SlotLength
+parseSlotDuration = option (mkSlotLength <$> auto) $ mconcat [
+      long "slot-duration"
+    , value (mkSlotLength 5)
+    , help "The slot duration (seconds)"
+    ]
+  where
+    mkSlotLength :: Integer -> SlotLength
+    mkSlotLength = slotLengthFromMillisec . (* 1000)
+
+parseProtocol :: Parser Protocol
+parseProtocol = asum [
+      flag' BFT $ mconcat [
+          long "bft"
+        , help "Use the BFT consensus algorithm"
+        ]
+    , flag' Praos $ mconcat [
+          long "praos"
+        , help "Use the Praos consensus algorithm"
+        ]
+    , flag' MockPBFT $ mconcat [
+          long "mock-pbft"
+        , help "Use the Permissive BFT consensus algorithm using a mock ledger"
+        ]
+    , flag' RealPBFT $ mconcat [
+          long "real-pbft"
+        , help "Use the Permissive BFT consensus algorithm using the real ledger"
+        ]
+    ]
+
+parseNodeId :: Parser NodeId
+parseNodeId =
+    option (fmap CoreId auto) (
+            long "node-id"
+         <> short 'n'
+         <> metavar "NODE-ID"
+         <> help "The ID for this node"
+    )
+
+parseCoreNodeId :: Parser CoreNodeId
+parseCoreNodeId =
+    option (fmap CoreNodeId auto) (
+            long "core-node-id"
+         <> short 'n'
+         <> metavar "CORE-NODE-ID"
+         <> help "The ID of the core node to which this client is connected."
+    )
+
+parseNumCoreNodes :: Parser NumCoreNodes
+parseNumCoreNodes =
+    option (fmap NumCoreNodes auto) (
+            long "num-core-nodes"
+         <> short 'm'
+         <> metavar "NUM-CORE-NODES"
+         <> help "The number of core nodes"
+    )
