--- cardano-node/NodeLib.hs
@@ -1,8 +1,0 @@
-module NodeLib
-    ( someFunc
-    ) where
-
-import Cardano.Prelude
-
-someFunc :: IO ()
-someFunc = putStrLn "placeholder"
