+++ wallet-client/CLI.hs
@@ -8,57 +8,12 @@ module CLI
   , fromProtocol
   ) where
 
-import           Data.Foldable (asum)
-import           Data.Semigroup ((<>))
 import           Options.Applicative
 
-import           Ouroboros.Consensus.Demo
 import           Ouroboros.Consensus.Demo.Run
 import           Ouroboros.Consensus.NodeId (CoreNodeId (..))
-import           Ouroboros.Consensus.Util
 
-import qualified Test.Cardano.Chain.Genesis.Dummy as Dummy
-  
-
-{-------------------------------------------------------------------------------
-  Untyped/typed protocol boundary
--------------------------------------------------------------------------------}
-
-data Protocol =
-    BFT
-  | Praos
-  | MockPBFT
-  | RealPBFT
-
-data SomeProtocol where
-  SomeProtocol :: RunDemo blk => DemoProtocol blk -> SomeProtocol
-
-fromProtocol :: Protocol -> IO SomeProtocol
-fromProtocol BFT =
-    case runDemo p of
-      Dict -> return $ SomeProtocol p
-  where
-    p = DemoBFT defaultSecurityParam
-fromProtocol Praos =
-    case runDemo p of
-      Dict -> return $ SomeProtocol p
-  where
-    p = DemoPraos defaultDemoPraosParams
-fromProtocol MockPBFT =
-    case runDemo p of
-      Dict -> return $ SomeProtocol p
-  where
-    p = DemoMockPBFT defaultDemoPBftParams
-fromProtocol RealPBFT =
-    case runDemo p of
-      Dict -> return $ SomeProtocol p
-  where
-    p = DemoRealPBFT defaultDemoPBftParams genesisConfig
-    genesisConfig = Dummy.dummyConfig
-
-{-------------------------------------------------------------------------------
-  Command line arguments
--------------------------------------------------------------------------------}
+import           Cardano.Node.CLI
 
 data CLI = CLI {
     cliCoreNodeId   :: CoreNodeId,
@@ -71,41 +26,3 @@ parseCLI = CLI
     <$> parseCoreNodeId
     <*> parseNumCoreNodes
     <*> parseProtocol
-
-parseProtocol :: Parser Protocol
-parseProtocol = asum [
-      flag' BFT $ mconcat [
-          long "bft"
-        , help "Use the BFT consensus algorithm"
-        ]
-    , flag' Praos $ mconcat [
-          long "praos"
-        , help "Use the Praos consensus algorithm"
-        ]
-    , flag' MockPBFT $ mconcat [
-          long "mock-pbft"
-        , help "Use the Permissive BFT consensus algorithm using a mock ledger"
-        ]
-    , flag' RealPBFT $ mconcat [
-          long "real-pbft"
-        , help "Use the Permissive BFT consensus algorithm using the real ledger"
-        ]
-    ]
-
-parseCoreNodeId :: Parser CoreNodeId
-parseCoreNodeId =
-    option (fmap CoreNodeId auto) (
-            long "core-node-id"
-         <> short 'n'
-         <> metavar "CORE-NODE-ID"
-         <> help "The ID of the core node to which this client is connected."
-    )
-
-parseNumCoreNodes :: Parser NumCoreNodes
-parseNumCoreNodes =
-    option (fmap NumCoreNodes auto) (
-            long "num-core-nodes"
-         <> short 'm'
-         <> metavar "NUM-CORE-NODES"
-         <> help "The number of core nodes"
-    )
